+++ scene/resources/importer_mesh.h
@@ -61,7 +61,7 @@ class ImporterMesh : public Resource {
 		Vector<LOD> lods;
 		Ref<Material> material;
 		String name;
-		uint32_t flags = 0;
+		uint64_t flags = 0;
 
 		struct LODComparator {
 			_FORCE_INLINE_ bool operator()(const LOD &l, const LOD &r) const {
@@ -93,7 +93,7 @@ public:
 	int get_blend_shape_count() const;
 	String get_blend_shape_name(int p_blend_shape) const;
 
-	void add_surface(Mesh::PrimitiveType p_primitive, const Array &p_arrays, const TypedArray<Array> &p_blend_shapes = Array(), const Dictionary &p_lods = Dictionary(), const Ref<Material> &p_material = Ref<Material>(), const String &p_name = String(), const uint32_t p_flags = 0);
+	void add_surface(Mesh::PrimitiveType p_primitive, const Array &p_arrays, const TypedArray<Array> &p_blend_shapes = Array(), const Dictionary &p_lods = Dictionary(), const Ref<Material> &p_material = Ref<Material>(), const String &p_name = String(), const uint64_t p_flags = 0);
 	int get_surface_count() const;
 
 	void set_blend_shape_mode(Mesh::BlendShapeMode p_blend_shape_mode);
@@ -108,7 +108,7 @@ public:
 	Vector<int> get_surface_lod_indices(int p_surface, int p_lod) const;
 	float get_surface_lod_size(int p_surface, int p_lod) const;
 	Ref<Material> get_surface_material(int p_surface) const;
-	uint32_t get_surface_format(int p_surface) const;
+	uint64_t get_surface_format(int p_surface) const;
 
 	void set_surface_material(int p_surface, const Ref<Material> &p_material);
